Reorganize examples into examples package with index
Move examples under examples/ with descriptive module names, add
helpers and __init__.py, and examples.index listing runnable modules.

Use relative imports (from .helpers import run_app); run as
python -m examples.<name> from repo root.

Resolve timekeys.txt via repo root path in timeline_multiaxis.

Update README with install and run instructions (under 20 lines of
shell in Reproduce).

Implements https://github.com/vecnode/dearpygui-plots/issues/2

Made-with: Cursor
