+++ examples/__init__.py
@@ -1,0 +1,5 @@
+"""Collection of DearPyGui plot examples.
+
+Run from the repository root, for example: python -m examples.bar_series
+See examples.index for a full list.
+"""
