+++ examples/helpers.py
